将ID改为int64格式

+++ server/commen/model.go
@@ -9,7 +9,7 @@ import (
 // 公用数据结构
 
 type GVA_MODEL struct {
-	ID        uint `gorm:"primarykey"`
+	ID        int64 `gorm:"primarykey"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	DeletedAt gorm.DeletedAt `gorm:"index" json:"-"`
